--- src/state/appHookedSelectors.ts
@@ -1,13 +1,0 @@
-import { hookedSelectors } from '@/lib/hookify/hookedSelectors';
-
-import { AppDispatch, RootState, store } from './_store';
-import { appQueryClient } from './appQueryClient';
-
-export const {
-  useQueryHf,
-  useHookedSelectorHf,
-  useHookedSelector,
-  useDispatchHf,
-  useAppSelectorHf,
-  createHookedSelector,
-} = hookedSelectors<RootState, AppDispatch>(store, appQueryClient);
